requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.2.0
rembg>=2.0.50
openai>=1.12.0
pathlib
typing
concurrent.futures
argparse
